Consolidated git* libraries
