+++ libraries/git/git_distributions.groovy
@@ -1,19 +1,21 @@
 /*
-  validate library configuration
+  Copyright © 2018 Booz Allen Hamilton. All Rights Reserved.
+  This software package is licensed under the Booz Allen Public License. The license can be found in the License file or at http://boozallen.github.io/licenses/bapl
+*/
+
+/*
+  Validate library configuration
 */
 @Validate
 void call(context){
+    String distributionConfig = config.distribution
     List options = [ "github", "github_enterprise", "gitlab" ]
-    Map distributionConfig = config.subMap(options)
 
-    // ensure only one distribution is configured
-    List configured_distributions = distributionConfig.keySet().toList()
-    if(configured_distributions.size() > 1){
-        error "You can only specify one distribution among ${options}, currently: ${configured_distributions}"
+    if (!options.contains(distributionConfig)) {
+        error "Distribution can only be set to one of the following: github, github_enterprise, gitlab. Currently: ${distributionConfig}"
     }
 
-    String distribution = distributionConfig.keySet().first()
-    env.GIT_LIBRARY_DISTRUBITION = distribution
+    env.GIT_LIBRARY_DISTRUBITION = distributionConfig
     def dist = this.fetch()
     dist.validate_configuration()
 }
